requirements.txt
openai
python-dotenv
pandas
openpyxl
tqdm
